Changed type int to uint16_t for port fields

+++ src/utils/Config.cc
@@ -33,13 +33,13 @@ uint32_t Config::RDMA_NUMAREGION = 0;
 std::string Config::RDMA_DEVICE_FILE_PATH;
 uint32_t Config::RDMA_IBPORT = 1;
 std::string Config::RDMA_SERVER_ADDRESSES = "172.18.94.20"; // ip node02 RDMA_INTERFACEs
-uint32_t Config::RDMA_PORT = 5200;
+uint16_t Config::RDMA_PORT = 5200;
 uint32_t Config::RDMA_MAX_WR = 4096;
 
 uint32_t Config::RDMA_UD_MTU = 4096;
 
 std::string Config::SEQUENCER_IP = "192.168.94.22"; //node02
-uint32_t Config::SEQUENCER_PORT = 5600;
+uint16_t Config::SEQUENCER_PORT = 5600;
 
 std::string Config::RDMA_INTERFACE = "ib0";
